# loan-validation/src/main/gen/api/openapi.yaml
openapi: 3.0.3
info:
  title: Loan Validation API
  version: 1.0.0
servers:
- url: /
paths:
  /loan-validations:
    post:
      operationId: validateLoan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanValidationRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanValidationResult'
          description: Resultado de validación
      summary: Valida un préstamo
      tags:
      - loan-validations
      x-content-type: application/json
      x-accepts:
      - application/json
components:
  schemas:
    LoanValidationRequest:
      example:
        monthlySalary: 0.09008281904610115
        termMonths: 6
        requestedAmount: 0.6127456183070403
        lastLoanDate: 2000-01-23
      properties:
        monthlySalary:
          format: double
          minimum: 0.01
          type: number
        requestedAmount:
          format: double
          minimum: 0.01
          type: number
        termMonths:
          maximum: 36
          minimum: 1
          type: integer
        lastLoanDate:
          description: Fecha del último préstamo del solicitante.
          format: date
          nullable: true
          type: string
      required:
      - monthlySalary
      - requestedAmount
      - termMonths
      type: object
    LoanValidationResult:
      example:
        reasons:
        - HAS_RECENT_LOANS
        - HAS_RECENT_LOANS
        monthlyPayment: 0.8008281904610115
        eligible: true
      properties:
        eligible:
          type: boolean
        reasons:
          items:
            enum:
            - HAS_RECENT_LOANS
            - PLAZO_MAXIMO_SUPERADO
            - CAPACIDAD_INSUFICIENTE
            - DATOS_INVALIDOS
            type: string
          type: array
        monthlyPayment:
          format: double
          type: number
      required:
      - eligible
      - monthlyPayment
      - reasons
      type: object