# loan-validation/target/generated-sources/openapi/src/main/java/com/techgirls/loanvalidation/api/LoanValidationsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.techgirls.loanvalidation.api;

import com.techgirls.loanvalidation.model.LoanValidationRequest;
import com.techgirls.loanvalidation.model.LoanValidationResult;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-09-10T18:11:06.484388400-05:00[America/Lima]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "loan-validations", description = "the loan-validations API")
public interface LoanValidationsApi {

    /**
     * POST /loan-validations : Valida un préstamo
     *
     * @param loanValidationRequest  (required)
     * @return Resultado de validación (status code 200)
     */
    @Operation(
        operationId = "validateLoan",
        summary = "Valida un préstamo",
        tags = { "loan-validations" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Resultado de validación", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LoanValidationResult.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/loan-validations",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<LoanValidationResult>> validateLoan(
        @Parameter(name = "LoanValidationRequest", description = "", required = true) @Valid @RequestBody Mono<LoanValidationRequest> loanValidationRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        for (MediaType mediaType : exchange.getRequest().getHeaders().getAccept()) {
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"reasons\" : [ \"HAS_RECENT_LOANS\", \"HAS_RECENT_LOANS\" ], \"monthlyPayment\" : 0.8008281904610115, \"eligible\" : true }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
        }
        return result.then(loanValidationRequest).then(Mono.empty());

    }

}